Remove simple demo.
