+++ app/src/main/java/com/novoda/downloadmanager/demo/MainActivity.java
@@ -13,6 +13,8 @@ import com.novoda.downloadmanager.DownloadBatchId;
 import com.novoda.downloadmanager.DownloadBatchIdCreator;
 import com.novoda.downloadmanager.DownloadBatchStatus;
 import com.novoda.downloadmanager.LiteDownloadManagerCommands;
+import com.novoda.downloadmanager.MigrationFactory;
+import com.novoda.downloadmanager.Migrator;
 import com.novoda.notils.logger.simple.Log;
 
 import java.util.List;
@@ -38,6 +40,25 @@ public class MainActivity extends AppCompatActivity {
 
         textViewBatch1 = findViewById(R.id.batch_1);
         textViewBatch2 = findViewById(R.id.batch_2);
+
+        View buttonMigrate = findViewById(R.id.button_migrate);
+        buttonMigrate.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                new Thread(new Runnable() {
+                    @Override
+                    public void run() {
+                        Migrator migrator = MigrationFactory.createVersionOneToVersionTwoMigrator(getApplicationContext(), getDatabasePath("downloads.db"), new Migrator.Callback() {
+                            @Override
+                            public void onMigrationComplete() {
+                                deleteDatabase("downloads.db");
+                            }
+                        });
+                        migrator.migrate();
+                    }
+                }).start();
+            }
+        });
 
         View buttonDownload = findViewById(R.id.button_start_downloading);
         buttonDownload.setOnClickListener(new View.OnClickListener() {
@@ -76,7 +97,7 @@ public class MainActivity extends AppCompatActivity {
                     public void onReceived(List<DownloadBatchStatus> downloadBatchStatuses) {
                         StringBuilder displayMessage = new StringBuilder();
                         Log.d("MainActivity", "List Size: " + downloadBatchStatuses.size());
-                        for(DownloadBatchStatus status : downloadBatchStatuses) {
+                        for (DownloadBatchStatus status : downloadBatchStatuses) {
                             if (status.bytesDownloaded() != status.bytesTotalSize()) {
                                 Log.d("MainActivity", status.getDownloadBatchTitle() + " not downloaded!");
                                 return;
@@ -89,7 +110,8 @@ public class MainActivity extends AppCompatActivity {
                         }
                         downloadedTitles.setText(displayMessage.toString());
                     }
-                });}
+                });
+            }
         });
 
         DemoApplication demoApplication = (DemoApplication) getApplicationContext();
