--- demo-simple/build.gradle
@@ -1,34 +1,0 @@
-buildscript {
-    repositories {
-        jcenter()
-    }
-    dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.0'
-    }
-}
-apply plugin: 'com.android.application'
-
-android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.0"
-
-    defaultConfig {
-        applicationId "com.novoda.downloadmanager.demo.simple"
-        minSdkVersion 14
-        targetSdkVersion 27
-        versionCode 1
-        versionName "1.0"
-
-        manifestPlaceholders = [downloadAuthority: "${applicationId}"]
-    }
-
-    lintOptions {
-        abortOnError false
-    }
-}
-
-dependencies {
-    compile project(':library')
-    compile 'com.android.support:appcompat-v7:27.0.0'
-    compile 'com.android.support:recyclerview-v7:27.0.0'
-}
