--- demo-simple/src/main/java/com/novoda/downloadmanager/demo/simple/MainActivity.java
@@ -1,112 +1,0 @@
-package com.novoda.downloadmanager.demo.simple;
-
-import android.os.Bundle;
-import android.os.Handler;
-import android.os.Looper;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.view.View;
-
-import com.facebook.stetho.Stetho;
-import com.novoda.downloadmanager.AllBatchStatusesCallback;
-import com.novoda.downloadmanager.Batch;
-import com.novoda.downloadmanager.DownloadBatchId;
-import com.novoda.downloadmanager.DownloadBatchIdCreator;
-import com.novoda.downloadmanager.DownloadBatchStatus;
-import com.novoda.downloadmanager.DownloadManagerBuilder;
-import com.novoda.downloadmanager.LiteDownloadManagerCommands;
-import com.novoda.downloadmanager.MigrationFactory;
-import com.novoda.downloadmanager.Migrator;
-import com.novoda.downloadmanager.demo.R;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class MainActivity extends AppCompatActivity {
-
-    private static final String BIG_FILE = "http://ipv4.download.thinkbroadband.com/200MB.zip";
-    private static final DownloadBatchId BEARD_ID = DownloadBatchIdCreator.createFrom("beard_id");
-
-    private LiteDownloadManagerCommands downloadManagerCommands;
-    private RecyclerView recyclerView;
-    private View emptyView;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        Stetho.initializeWithDefaults(this);
-        setContentView(R.layout.activity_main);
-        emptyView = findViewById(R.id.main_no_downloads_view);
-        recyclerView = findViewById(R.id.main_downloads_list);
-        recyclerView.setLayoutManager(new LinearLayoutManager(this));
-        Handler handler = new Handler(Looper.getMainLooper());
-        downloadManagerCommands = DownloadManagerBuilder
-                .newInstance(this, handler, R.mipmap.ic_launcher)
-                .build();
-
-        setupDownloadingExample();
-        setupQueryingExample();
-    }
-
-    private void setupDownloadingExample() {
-        findViewById(R.id.main_download_button).setOnClickListener(
-                new View.OnClickListener() {
-                    @Override
-                    public void onClick(@NonNull View v) {
-                        final Batch batch = new Batch.Builder(BEARD_ID, "Family of Penguins")
-                                .addFile(BIG_FILE)
-                                .build();
-                        downloadManagerCommands.download(batch);
-                    }
-                });
-    }
-
-    private void setupQueryingExample() {
-        findViewById(R.id.main_refresh_button).setOnClickListener(
-                new View.OnClickListener() {
-                    @Override
-                    public void onClick(@NonNull View v) {
-                        queryForDownloads();
-                        migrateV1DownloadsUsingNewThread();
-                    }
-                }
-        );
-    }
-
-    private void queryForDownloads() {
-        downloadManagerCommands.getAllDownloadBatchStatuses(new AllBatchStatusesCallback() {
-            @Override
-            public void onReceived(List<DownloadBatchStatus> downloadBatchStatuses) {
-                List<BeardDownload> beardDownloads = new ArrayList<>(downloadBatchStatuses.size());
-                for (DownloadBatchStatus downloadBatchStatus : downloadBatchStatuses) {
-                    BeardDownload beardDownload = new BeardDownload(downloadBatchStatus.getDownloadBatchTitle(), downloadBatchStatus.status());
-                    beardDownloads.add(beardDownload);
-                }
-                onQueryResult(beardDownloads);
-            }
-        });
-    }
-
-    public void onQueryResult(List<BeardDownload> beardDownloads) {
-        recyclerView.setAdapter(new BeardDownloadAdapter(beardDownloads));
-        emptyView.setVisibility(beardDownloads.isEmpty() ? View.VISIBLE : View.GONE);
-    }
-
-    private void migrateV1DownloadsUsingNewThread() {
-        new Thread(new Runnable() {
-            @Override
-            public void run() {
-                Migrator migrator = MigrationFactory.createVersionOneToVersionTwoMigrator(getApplicationContext(), getDatabasePath("downloads.db"), new Migrator.Callback() {
-                    @Override
-                    public void onMigrationComplete() {
-                        deleteDatabase("downloads.db");
-                    }
-                });
-                migrator.migrate();
-            }
-        }).start();
-    }
-
-}
