--- demo-simple/src/main/res/layout/activity_main.xml
@@ -1,35 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:tools="http://schemas.android.com/tools"
-  android:layout_width="match_parent"
-  android:layout_height="match_parent"
-  android:orientation="vertical"
-  android:paddingBottom="@dimen/activity_vertical_margin"
-  android:paddingLeft="@dimen/activity_horizontal_margin"
-  android:paddingRight="@dimen/activity_horizontal_margin"
-  android:paddingTop="@dimen/activity_vertical_margin"
-  tools:context=".MainActivity">
-
-  <Button
-    android:id="@+id/main_download_button"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:text="@string/button_download_text" />
-
-  <Button
-    android:id="@+id/main_refresh_button"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:text="@string/button_refresh_text" />
-
-  <TextView
-    android:id="@+id/main_no_downloads_view"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:text="@string/no_downloads" />
-
-  <android.support.v7.widget.RecyclerView
-    android:id="@+id/main_downloads_list"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content" />
-
-</LinearLayout>
