--- demo-simple/src/main/res/values/integers.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-  <integer name="max_concurrent_downloads">5</integer>
-</resources>
