--- demo-simple/src/main/res/values/strings.xml
@@ -1,7 +1,0 @@
-<resources>
-  <string name="app_name">Beard Downloader</string>
-
-  <string name="no_downloads">No downloads yet.\nTap download and then refresh to query the download manager!</string>
-  <string name="button_refresh_text">Refresh</string>
-  <string name="button_download_text">Download a podcast</string>
-</resources>
