--- demo-simple/src/main/res/values/styles.xml
@@ -1,8 +1,0 @@
-<resources>
-
-  <!-- Base application theme. -->
-  <style name="AppTheme" parent="Theme.AppCompat.Light.DarkActionBar">
-    <!-- Customize your theme here. -->
-  </style>
-
-</resources>
